// src/tlt-rms/dto/create-tlt-rm.dto.ts
export class CreateTltRmDto {
    pgNonPg: string;

  warehouse: string;

  documentNo: string;

  documentDate: string;

  vendorName: string;

  itemName: string;

  unit: string;

  receivedPieces: number;

  receivedQuantity: number;

  issuedPieces: number;

  issuedQuantity: number;

  balancePieces: number;

  balanceQuantity: number;

  balanceAmount: number;

  pendingDays: number;

  maxLength: number;

  minLength: number;

  averageLength: number;

  length: number;

  width: number;

  costCenter: string;

  section: string;

  createdBy: string;
  
  itemId: string;
}